requirements.txt
# ROOTY! CLOUD REQUIREMENTS (OPTIMIZED)
streamlit==1.56.0
gspread==5.12.4
google-auth==2.49.2
google-cloud-firestore==2.27.0
pandas==3.0.2
requests==2.33.1
